Web controllers now use REST Controllers; added views for individual schema entries

+++ src/main/java/cookcollab/com/cookcollab/data/entity/Rating.java
@@ -54,7 +54,7 @@ public class Rating {
 		this.userID = userID;
 	}
 
-	public int getRating() {
+	public long getRating() {
 		return this.rating;
 	}
